added proto script & updated gitignore
