+++ .gitignore
@@ -1,6 +1,21 @@
+# node & SvelteKit
 node_modules
+
+# Go 
 dist
+
+# dev data 
 mongo
+
+# git
 .gitcredentials
+
+# macOS 
 .DS_Store
+
+# ???
 .env
+
+# httpBin
+binManagerApp
+apiGateWayApp
