+++ httpBin.code-workspace
@@ -14,7 +14,7 @@
     }
   ],
   "settings": {
-    "workbench.colorTheme": "Svelte Dark",
+    "workbench.colorTheme": "Atom One Dark",
     "theme-by-language.useGlobalConfigOnly": false
   }
 }
